ready-made parsers
