+++ basic.go
@@ -78,21 +78,6 @@ func Not(str string) Combinator[string] {
 	}
 }
 
-// Crlf must match either a CR or CRLF line ending.
-//
-//	chomp.Crlf()("\r\nHello")
-//	// ("Hello", "\r\n", nil)
-func Crlf() Combinator[string] {
-	return func(s string) (string, string, error) {
-		idx := strings.Index(s, "\n")
-		if idx == 0 || (idx == 1 && s[0] == '\r') {
-			return s[idx+1:], s[:idx+1], nil
-		}
-
-		return s, "", CombinatorParseError{Text: s, Type: "crlf"}
-	}
-}
-
 // OneOf must match a single character at the beginning of the text from the
 // provided sequence.
 //
@@ -147,60 +132,5 @@ func Until(str string) Combinator[string] {
 		}
 
 		return s, "", CombinatorParseError{Input: str, Text: s, Type: "until"}
-	}
-}
-
-// Prefixed will firstly scan the input text for a defined prefix and discard it.
-// The remaining input text will be matched against the [Combinator] and returned
-// if successful. Both combinators must match.
-//
-//	chomp.Prefixed(
-//		chomp.Tag("Hello"),
-//		chomp.Tag(`"`))(`"Hello, World!"`)
-//	// (`, World!"`, "Hello", nil)
-func Prefixed(c, pre Combinator[string]) Combinator[string] {
-	return func(s string) (string, string, error) {
-		rem, _, err := pre(s)
-		if err != nil {
-			return rem, "", err
-		}
-
-		return c(rem)
-	}
-}
-
-// Suffixed will firstly scan the input text and match it against the [Combinator].
-// The remaining text will be scanned for a defined suffix and discarded. Both
-// combinators must match.
-//
-//	chomp.Suffixed(
-//		chomp.Tag("Hello"),
-//		chomp.Tag(", "))("Hello, World!")
-//	// ("World!", "Hello", nil)
-func Suffixed(c, suf Combinator[string]) Combinator[string] {
-	return func(s string) (string, string, error) {
-		rem, ext, err := c(s)
-		if err != nil {
-			return rem, "", err
-		}
-
-		rem, _, err = suf(rem)
-		if err != nil {
-			return rem, "", err
-		}
-
-		return rem, ext, nil
-	}
-}
-
-// Eol will scan the text until it encounters any ASCII line ending characters
-// identified by the [IsLineEnding] predicate. All text before the line ending
-// will be returned. The line ending, if detected, will be discarded.
-//
-//	chomp.Eol()(`Hello, World!\nIt's a great day!`)
-//	// ("It's a great day!", "Hello, World!", nil)
-func Eol() Combinator[string] {
-	return func(s string) (string, string, error) {
-		return Suffixed(WhileNotN(IsLineEnding, 0), Opt(Crlf()))(s)
 	}
 }
