+++ basic_test.go
@@ -157,47 +157,6 @@ func TestUntil(t *testing.T) {
 	}
 }
 
-func TestCrlf(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name  string
-		input string
-		rem   string
-		ext   string
-	}{
-		{
-			name:  "LF",
-			input: "\nHello",
-			rem:   "Hello",
-			ext:   "\n",
-		},
-		{
-			name:  "CRLF",
-			input: "\r\nこんにちは",
-			rem:   "こんにちは",
-			ext:   "\r\n",
-		},
-		{
-			name:  "LFOnly",
-			input: "\n",
-			rem:   "",
-			ext:   "\n",
-		},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			rem, ext, err := chomp.Crlf()(tt.input)
-
-			require.NoError(t, err)
-			assert.Equal(t, tt.rem, rem)
-			assert.Equal(t, tt.ext, ext)
-		})
-	}
-}
-
 func TestOneOf(t *testing.T) {
 	t.Parallel()
 
@@ -274,26 +233,6 @@ func TestNoneOf(t *testing.T) {
 	}
 }
 
-func TestPrefixed(t *testing.T) {
-	t.Parallel()
-
-	rem, ext, err := chomp.Prefixed(chomp.Tag("Hello"), chomp.Tag(`"`))(`"Hello, World"`)
-
-	require.NoError(t, err)
-	assert.Equal(t, `, World"`, rem)
-	assert.Equal(t, "Hello", ext)
-}
-
-func TestSuffixed(t *testing.T) {
-	t.Parallel()
-
-	rem, ext, err := chomp.Suffixed(chomp.Tag("Hello"), chomp.Tag(","))("Hello, World")
-
-	require.NoError(t, err)
-	assert.Equal(t, " World", rem)
-	assert.Equal(t, "Hello", ext)
-}
-
 func TestCombinatorError(t *testing.T) {
 	t.Parallel()
 
@@ -311,47 +250,4 @@ func TestParserCombinatorError(t *testing.T) {
 		chomp.Tag("marvel"))("the legend of batman:dc:9781801260336:£19.99")
 
 	assert.EqualError(t, err, "(all) parser failed. (tag) combinator failed to parse text 'dc:9781801260336:£19.99' with input 'marvel'")
-}
-
-func TestEol(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name  string
-		input string
-		rem   string
-		ext   string
-	}{
-		{
-			name: "LF",
-			input: `Hello, World!
-It's a great day`,
-			rem: "It's a great day",
-			ext: "Hello, World!",
-		},
-		{
-			name:  "NoLF",
-			input: "こんにちは、おはよう",
-			rem:   "",
-			ext:   "こんにちは、おはよう",
-		},
-		{
-			name: "EmptyLineBeforeLF",
-			input: `
-こんにちは、おはよう`,
-			rem: "こんにちは、おはよう",
-			ext: "",
-		},
-	}
-	for _, tt := range tests {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-			rem, ext, err := chomp.Eol()(tt.input)
-
-			require.NoError(t, err)
-			assert.Equal(t, tt.rem, rem)
-			assert.Equal(t, tt.ext, ext)
-		})
-	}
 }
